src/requirements.txt
Adafruit-Blinka==6.6.0
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-rgb-display==3.10.6
Adafruit-PlatformDetect==3.8.0
Adafruit-PureIO==1.1.8
certifi==2020.12.5
chardet==4.0.0
colorzero==2.0
gpiozero==1.6.2
idna==2.10
Pillow==8.2.0
pyftdi==0.52.9
pyserial==3.5
pyusb==1.1.1
requests==2.25.1
rpi-ws281x==4.2.6
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
urllib3==1.26.4